feat: "horario post"

+++ equipos/equipo-03/backend/src/main/java/com/canchavoley/backend/repository/HorarioRepository.java
@@ -3,6 +3,5 @@ package com.canchavoley.backend.repository;
 import com.canchavoley.backend.model.Horario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HorarioRepository extends JpaRepository<Horario, Long> {
-
+public interface HorarioRepository extends JpaRepository<Horario, Integer> {
 }
